+++ src/test/kotlin/kotools/types/string/NotBlankStringTest.kt
@@ -1,6 +1,7 @@
 package kotools.types.string
 
 import kotools.assert.*
+import kotools.types.number.NonZeroInt
 import kotools.types.number.PositiveInt
 
 class NotBlankStringTest {
@@ -248,6 +249,92 @@ class NotBlankStringTest {
     }
 
     // ---------- Conversions ----------
+
+    @Nested
+    inner class ToNonZeroInt {
+        @Test
+        fun `should return a non-zero int with a string representation of 1`() {
+            // GIVEN
+            val value = 1
+            val string = NotBlankString(value.toString())
+            // WHEN
+            val result: NonZeroInt = assertPass(string::toNonZeroInt)
+            // THEN
+            result.value assertEquals value
+        }
+
+        @Test
+        fun `should return a non-zero int with a string representation of -1`() {
+            // GIVEN
+            val value = -1
+            val string = NotBlankString(value.toString())
+            // WHEN
+            val result: NonZeroInt = assertPass(string::toNonZeroInt)
+            // THEN
+            result.value assertEquals value
+        }
+
+        @Test
+        fun `should throw an error with an invalid string`() {
+            // GIVEN
+            val string = NotBlankString("hello")
+            // WHEN & THEN
+            assertFailsWith<NumberFormatException>(string::toNonZeroInt)
+        }
+
+        @Test
+        fun `should throw an error with a string representation of 0`() {
+            // GIVEN
+            val string = NotBlankString("0")
+            // WHEN & THEN
+            assertFailsWith<IllegalArgumentException>(string::toNonZeroInt)
+        }
+    }
+
+    @Nested
+    inner class ToNonZeroIntOrNull {
+        @Test
+        fun `should return a non-zero int with a string representation of 1`() {
+            // GIVEN
+            val value = 1
+            val string = NotBlankString(value.toString())
+            // WHEN
+            val result: NonZeroInt? = string.toNonZeroIntOrNull()
+            // THEN
+            result.assertNotNull().value assertEquals value
+        }
+
+        @Test
+        fun `should return a non-zero int with a string representation of -1`() {
+            // GIVEN
+            val value = -1
+            val string = NotBlankString(value.toString())
+            // WHEN
+            val result: NonZeroInt? = string.toNonZeroIntOrNull()
+            // THEN
+            result.assertNotNull().value assertEquals value
+        }
+
+        @Test
+        fun `should return null with an invalid string`() {
+            // GIVEN
+            val string = NotBlankString("hello")
+            // WHEN
+            val result: NonZeroInt? = string.toNonZeroIntOrNull()
+            // THEN
+            result.assertNull()
+        }
+
+        @Test
+        fun `should return null with a string representation of 0`() {
+            // GIVEN
+            val string = NotBlankString("0")
+            // WHEN
+            val result: NonZeroInt? = string.toNonZeroIntOrNull()
+            // THEN
+            result.assertNull()
+        }
+    }
 
     @Nested
     inner class ToString {