+++ changelog.md
@@ -17,6 +17,11 @@ All notable changes to this project will be documented in this file.
 
 ## Work in progress
 
+### Added
+
+- Conversions from `String` and `NotBlankString` to `NonZeroInt`
+  ([#48](https://github.com/kotools/types/issues/48)).
+
 ## Version 2.0.0 - 2022/08/01
 
 ### Added